Hopefully fixes non-unicode issue inelegantly.
There is an issue in the IRC library dogexam uses which causes it to raise an exception by default if a non-unicode-decodable character is supplied by a user's IRC client (non-standard clients are stupid, I do not have a high opinion of how IRC works on this aspect). Unfortunately the IRC library does not have a mutator method to change the default behaviour to ignoring these characters conveniently. Due to how this bot was originally built, it requires some very inelegant hack to modify the class variable about two parent classes up. Therefore I have resorted to reinitialise the bot instance (and consequently restart the bot) when this very rare issue is encountered.